requirements.txt
flask==3.1.0
openai==1.58.1
python-dotenv